requirements.txt
matplotlib==3.2.2
pandas==1.1.5
torchvision==0.4.0
numpy==1.19.5
albumentations==0.4.6
torch==1.2.0
transformers==4.8.2
opencv_contrib_python==4.1.2.30
tqdm==4.41.1
Pillow==8.3.0
PyYAML==5.4.1
scikit_learn==0.24.2
sentencepiece
